adapt to api change of stock system

+++ org.dieschnittstelle.jee.esa.ejb.client/src/main/java/org/dieschnittstelle/jee/esa/ejb/client/ejbclients/StockSystemClient.java
@@ -53,7 +53,7 @@ public class StockSystemClient implements StockSystemRemote {
 	}
 
 	@Override
-	public List<Integer> getPointsOfSale(IndividualisedProductItem product) {
+	public List<Long> getPointsOfSale(IndividualisedProductItem product) {
 		//return this.proxy.getPointsOfSale(product);
 		return null;
 	}
